+++ src/boot/utl.js
@@ -1,5 +1,6 @@
 import UtlDialog from 'components/utl/utl-dialog'
 import UtlBreadcrumbs from 'components/utl/utl-breadcrumbs'
+import UtlExpansion from 'components/utl/utl-expansion'
 import MxmMarkdown from 'components/mxm-markdown'
 import { version } from '../../package.json'
 import map from 'lodash/map'
@@ -25,6 +26,7 @@ export default ({ app, Vue, router }) => {
 
   Vue.component('utl-dialog', UtlDialog)
   Vue.component('utl-breadcrumbs', UtlBreadcrumbs)
+  Vue.component('utl-expansion', UtlExpansion)
   Vue.component('mxm-markdown', MxmMarkdown)
 }