+++ src/components/asset-new-button.vue
@@ -38,6 +38,7 @@
         <div>
           Description:
           <mxm-markdown
+            expandable
             class="bg-light-blue-1"
             style="min-height:4em;min-width:24em"
             :text="description"
@@ -66,13 +67,11 @@
       }
     },
 
-    data() {
-      return {
-        dialogOpened: false,
-        assetId: '',
-        description: ''
-      }
-    },
+    data: () => ({
+      dialogOpened: false,
+      assetId: '',
+      description: null,
+    }),
 
     computed: {
       okToSubmit() {
@@ -87,7 +86,7 @@
     methods: {
       openDialog() {
         this.assetId = ''
-        this.description = ''
+        this.description = null
         this.dialogOpened = true
       },