+++ src/pages/Missions.vue
@@ -35,7 +35,7 @@
       </div>
 
       <q-td slot="body-cell-missionId" slot-scope="props" :props="props"
-            style="width:5px"
+            style="width:5px;vertical-align:top"
       >
         <router-link
           style="text-decoration:none"
@@ -46,7 +46,7 @@
       </q-td>
 
       <q-td slot="body-cell-missionTplId" slot-scope="props" :props="props"
-            style="width:5px"
+            style="width:5px;vertical-align:top"
       >
         <router-link
           style="text-decoration:none"
@@ -57,7 +57,7 @@
       </q-td>
 
       <q-td slot="body-cell-assetId" slot-scope="props" :props="props"
-            style="width:5px"
+            style="width:5px;vertical-align:top"
       >
         <router-link
           style="text-decoration:none"
@@ -69,11 +69,14 @@
 
       <q-td slot="body-cell-description" slot-scope="props" :props="props"
       >
-        <mxm-markdown simple hide-empty :text="props.value"/>
+        <mxm-markdown
+          expandable :expandable-subtitle-limit="50"
+          simple hide-empty :text="props.value"
+        />
       </q-td>
 
       <q-td slot="body-cell-missionStatus" slot-scope="props" :props="props"
-            style="width:5px"
+            style="width:5px;vertical-align:top"
       >
         <q-chip dense>{{props.value}}</q-chip>
       </q-td>